Add null validation

+++ backend/src/main/java/com/topolski/backend/controller/AdminController.java
@@ -9,6 +9,7 @@ import com.topolski.backend.service.ProductService;
 import com.topolski.backend.service.ReviewService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.apache.logging.log4j.util.Strings;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -44,9 +45,11 @@ public class AdminController {
     }
 
     @PutMapping("/products/{id}")
-    public ResponseEntity<ProductDTO> updateProduct(
+    public ResponseEntity<?> updateProduct(
             @PathVariable Long id,
             @RequestBody @Valid ProductRequest productRequest) {
+        if(id == null) return ResponseEntity.badRequest().body(new ServerResponse("Id cannot be null"));
+
         ProductDTO updatedProduct = productService.updateProduct(id, productRequest);
         return ResponseEntity.ok(updatedProduct);
     }
@@ -54,12 +57,16 @@ public class AdminController {
     @PostMapping("/products/{id}/image")
     public ResponseEntity<ServerResponse> uploadFile(@PathVariable Long id,
                                                      @RequestParam("file") MultipartFile file) {
+        if(id == null || file == null) return ResponseEntity.badRequest().body(new ServerResponse("Id or imageUrl file are not correctly configured"));
+
         productService.addProductImageUrl(id, file.getName());
         return ResponseEntity.ok().body(new ServerResponse(s3Service.putObject(file, file.getOriginalFilename())));
     }
 
     @DeleteMapping("/products/{id}/image")
     public ResponseEntity<ServerResponse> deleteFile(@PathVariable Long id, @RequestParam("imageUrl") String imageUrl) {
+        if(id == null || Strings.isBlank(imageUrl)) return ResponseEntity.badRequest().body(new ServerResponse("Id or imageUrl file are not correctly configured"));
+
         productService.removeProductImageUrl(id, imageUrl);
         return ResponseEntity.ok().body(new ServerResponse(s3Service.deleteObject(imageUrl)));
     }
@@ -71,6 +78,8 @@ public class AdminController {
 
     @DeleteMapping("/reviews/{id}")
     public ResponseEntity<ServerResponse> deleteReview(@PathVariable Long id) {
+        if(id == null) return ResponseEntity.badRequest().body(new ServerResponse("Id cannot be null"));
+
         reviewService.deleteReviewById(id);
         return ResponseEntity.ok(new ServerResponse("Review deleted successfully"));
     }
